upgraded into karaf 2.2.9, cxf 2.6.2 to be compatible with dyrealla 0.7.0-SNAPSHOT

+++ coined/src/test/java/pl/cyfronet/coin/impl/CoinedAsserts.java
@@ -48,7 +48,7 @@ public class CoinedAsserts {
 		assertEquals(type, asEndpoint.getType());
 		assertEquals(atEndpoint.getInvocation_path(),
 				asEndpoint.getInvocationPath());
-		assertEquals(atEndpoint.getPort().intValue(), asEndpoint.getPort());
+		assertEquals(atEndpoint.getPort(), asEndpoint.getPort());
 		assertEquals(atEndpoint.getService_name(), asEndpoint.getServiceName());
 	}
 }
